Force singlet states and order orbitals according to decreasing occupations
Sometimes the FCI in natural orbital basis produces a different energy
than the FCI in the SCF orbital basis. Maybe this is due to the fact
that originally the natural orbitals were ordered according to increasing
occupations. The latter might mean that the FCI guess would start at
the highest energy Slater determinant instead of the lowest energy
one. Possibly the code gets stuck on a non-ground state wavefunction.